Day 23 Sonar tweaks

+++ src/main/java/com/ikueb/advent17/Day23.java
@@ -5,7 +5,7 @@ import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.stream.IntStream;
+import java.util.stream.LongStream;
 
 final class Day23 {
 
@@ -31,7 +31,7 @@ final class Day23 {
     static long getH(int value) {
         // with lots of help from reddit
         int seed = value * 100 + 100000;
-        return IntStream.iterate(0, i -> i <= 1000, i -> i + 1)
+        return LongStream.iterate(0, i -> i <= 1000, i -> i + 1)
                 .filter(i -> !BigInteger.valueOf(seed + 17 * i).isProbablePrime(100))
                 .count();
     }
